refactor: split module JS into individual files; bundle during build
This improves the readability of the code, while eliminating any
potential performance negatives in the final build.
