+++ module/js/ModuleSettingConsts.js
@@ -1,0 +1,6 @@
+export class ModuleSettingConsts {
+	static MENU_CONFIGURE_DEPENDENCIES = "menuConfigureDependencies";
+	static IS_NOTIFY_ON_LOAD = "isNotifyOnLoad";
+	static MENU_CONFIGURE_OPTIONAL_DEPENDENCIES = "menuConfigureOptionalDependencies";
+	static OPTIONAL_DEPENDENCY_NOTIFICATION_CONFIG = "optionalDependencyNotificationConfig";
+}
