+++ module/js/Util.js
@@ -1,0 +1,7 @@
+export class Util {
+	static LGT = [
+		`%cPlutonium Addon: Data`,
+		`color: #5494cc; font-weight: bold;`,
+		`|`,
+	];
+}
