+++ module/shared/SharedConsts.js
@@ -2,8 +2,6 @@ class SharedConsts {
 	static MODULE_ID = `plutonium-addon-automation`;
 	static MODULE_TITLE = "Plutonium Addon: Automation";
 
-	static MODULE_DIR = `./dist/${SharedConsts.MODULE_ID}`;
-
 	static MODULE_PATH = `modules/${SharedConsts.MODULE_ID}`;
 }
