+++ script/build.js
@@ -1,3 +1,11 @@
 import {buildTask} from "./build-task.js";
+import {Command} from "commander";
 
-buildTask().then(null);
+const program = new Command()
+	.option("--dev", `If this is a "dev" build`)
+;
+
+program.parse(process.argv);
+const params = program.opts();
+
+buildTask({isBundle: !params.dev}).then(null);
